requirements.txt
# Core LangChain dependencies
langchain>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
langchain-community>=0.3.0

# Google Gemini API
google-generativeai>=0.8.0

# Web search
tavily-python>=0.5.0

# Environment and configuration
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0

# Property-based testing
hypothesis>=6.0.0

# Video processing
ffmpeg-python>=0.2.0

# HTTP requests (for TAMUS API)
requests>=2.31.0

# Additional utilities
typing-extensions>=4.0.0
